+++ components/cms/_form.php
@@ -65,20 +65,6 @@ foreach (\Yii::$app->cms->emailTemplates as $code => $data)
 
 <?= $form->fieldSetEnd(); ?>
 
-<?= $form->fieldSet('Агенты'); ?>
-    <? \yii\bootstrap\Alert::begin([
-        'options' => [
-          'class' => 'alert-warning',
-      ],
-    ]); ?>
-    <b>Внимание!</b> По возможности, используйте работу агентов на хитах, это увеличит производительность вашего сайта.
-    <? \yii\bootstrap\Alert::end()?>
-
-    <?= $form->fieldRadioListBoolean($model, 'enabledHitAgents')->hint('Если вы отключаете выполнение агентов на хитах, то не забудте включить их в задание cron'); ?>
-    <?= $form->fieldInputInt($model, 'hitAgentsInterval')->hint('Если агенты выполняются на хитах, то их выполнение будет осуществляться с заданным переидом (сек.)'); ?>
-<?= $form->fieldSetEnd(); ?>
-
-
 
 <?= $form->fieldSet('Безопасность'); ?>
     <?= $form->fieldInputInt($model, 'passwordResetTokenExpire')->hint('Другими словами, ссылки на восстановление пароля перестанут работать через указанное время'); ?>
@@ -106,41 +92,6 @@ foreach (\Yii::$app->cms->emailTemplates as $code => $data)
     )->hint('Так же после созданию пользователя, ему будут назначены, выбранные группы.'); ?>
 
 <?= $form->fieldSetEnd(); ?>
-
-<?/*= $form->fieldSet('Отладка'); */?><!--
-    <?/* \yii\bootstrap\Alert::begin([
-        'options' => [
-          'class' => 'alert-warning',
-      ],
-    ]); */?>
-    <b>Внимание!</b> Помните, в режиме отладки ваш сайт работает медленнее. Не включайте отладку на рабочих сайтах.
-    <?/* \yii\bootstrap\Alert::end()*/?>
-    <?/*= $form->fieldRadioListBoolean($model, 'debugEnabled'); */?>
-    <?/*= $form->field($model, 'debugAllowedIPs')->textarea([
-        'placeholder' => '80.243.13.242,127.*'
-    ])->hint('Укажите ip адреса для которых будет показана отладочная панель дебага через запятую.'); */?>
-    <p><b>Ваш ip:</b> <?/*= \Yii::$app->getRequest()->getUserIP(); */?></p>
-
-
---><?/*= $form->fieldSetEnd(); */?>
-
-<?/*= $form->fieldSet('Разработка'); */?><!--
-    <?/* \yii\bootstrap\Alert::begin([
-        'options' => [
-          'class' => 'alert-warning',
-      ],
-    ]); */?>
-    Обратите внимание на эти настройки
-    <?/* \yii\bootstrap\Alert::end()*/?>
-    <?/*= $form->fieldRadioListBoolean($model, 'giiEnabled'); */?>
-    <?/*= $form->field($model, 'giiAllowedIPs')->textarea([
-        'placeholder' => '80.243.13.242,127.*'
-    ])->hint('Укажите ip адреса для которых будет включен генератор кода через запятую.'); */?>
-    <p><b>Ваш ip:</b> <?/*= \Yii::$app->getRequest()->getUserIP(); */?></p>
-
-
---><?/*= $form->fieldSetEnd(); */?>
-
 
 <?= $form->fieldSet('Доступ'); ?>