+++ components/marketplace/MarketplaceApi.php
@@ -80,7 +80,6 @@ class MarketplaceApi extends Component
 
             $data = array_merge($data, [
                 'sx-serverName' => \Yii::$app->request->serverName,
-                'sx-key'        => \Yii::$app->cms->licenseKey,
                 'sx-version'    => \Yii::$app->cms->extension->version,
                 'sx-email'      => \Yii::$app->cms->adminEmail,
             ]);